Ignore errors in shutils.rmtree
Causes shutils.rmtree() to ignore errors when attempting to clean up the
temporary configuration directory. This is necessary on Windows when
executing a program during Configuration.check() because it doesn't
appear to allow the parent directory to be removed if it recently had a
running executable file in it.

Without this change, this trivial wscript fails on Windows platforms:
    def options(opt):
        opt.load('compiler_c')

    def configure(cfg):
        cfg.load('compiler_c')
        cfg.check(execute=True, msg='Executing program')