Improve shutil.rmtree() for test executables
Windows systems have some internal services that like to open
executables when they are run and examine them (i.e. the Application
Experience Lookup Service). Normally, this isn't an issue because the
service closes the executable promptly. However, here we are trying to
execute a test executable which will run for a very short period of
time, then attempting to immediately delete the directory that contains
that executable. Both of these steps can happen before the Windows
services can release their handle on the executable, which will give us
an error saying the directory is not empty, so we can't delete it. All
we can do is wait for a short time to give the services a chance to
close their handles and try again